Ensure remove Tcp(Server|Client) from the loop

+++ src/cpp/scaler/ymq/tcp_client.cpp
@@ -72,6 +72,10 @@ void TcpClient::onRead()
 
 void TcpClient::onWrite()
 {
+    if (!_rawClient.nativeHandle()) {
+        return;
+    }
+
     if (_rawClient.needRetry()) {
         _rawClient.destroy();
         retry();
@@ -109,12 +113,17 @@ void TcpClient::retry()
     _retryIdentifier = _eventLoopThread->_eventLoop.executeAt(at, [this] { this->onCreated(); });
 }
 
-TcpClient::~TcpClient() noexcept
+void TcpClient::disconnect()
 {
     if (_rawClient.nativeHandle()) {
         _eventLoopThread->_eventLoop.removeFdFromLoop(_rawClient.nativeHandle());
         _rawClient.destroy();
     }
+}
+
+TcpClient::~TcpClient() noexcept
+{
+    disconnect();
     if (_retryTimes > 0) {
         _eventLoopThread->_eventLoop.cancelExecution(_retryIdentifier);
     }
